Create login form on image
The login form on an Image is used on many websites. Like hotels' website that contains pictures of the hotels or some organizations that organize special events holding that events picture and login form on that. In that case, you can design a login or registration form on that picture.